Added more synchronisation between controller Added comment to data_structures && config && constants. Fixed ordering bug in controller_handle

+++ common/src/data_structures/full_requests_matrix.rs
@@ -2,20 +2,24 @@ use crate::config::{CLIENT_COUNT, N_FLOOR, VALID_CLIENT_IDS};
 use crate::connection::event_handle::handle_state::ConnectionIdentifier;
 use crate::constants::{N_BUTTONS, N_HALL_BUTTONS};
 
-const FLOOR_ARRAY_SIZE: usize = N_HALL_BUTTONS + CLIENT_COUNT as usize;
+const FLOOR_ARRAY_SIZE: usize = N_HALL_BUTTONS + CLIENT_COUNT;
 const CLIENTS_RAW_SIZE: usize = N_FLOOR as usize * FLOOR_ARRAY_SIZE;
-pub(super) const FCRM_RAW_SIZE: usize = CLIENT_COUNT as usize + CLIENTS_RAW_SIZE;
+/// Size of the raw bytes array generated from encoding a [FullControllerRequestsMatrix].
+pub(super) const FCRM_RAW_SIZE: usize = CLIENT_COUNT + CLIENTS_RAW_SIZE;
 
 #[derive(Debug, Copy, Clone)]
+/// Utility struct used to facilitate passing the requests matrix from a controller to another one.
 pub struct FullControllerRequestsMatrix {
-    client_ids: [u8; CLIENT_COUNT as usize],
+    client_ids: [u8; CLIENT_COUNT],
     clients: [[bool; FLOOR_ARRAY_SIZE]; N_FLOOR as usize]
 }
 
 impl FullControllerRequestsMatrix {
+    /// Create a [FullControllerRequestsMatrix] from two merged array, representing respectively
+    /// all the hall requests and all the cab requests.
     pub fn from(
         merged_hall_requests: [[bool; N_HALL_BUTTONS]; N_FLOOR as usize],
-        clients_cab_requests: [[bool; CLIENT_COUNT as usize]; N_FLOOR as usize]
+        clients_cab_requests: [[bool; CLIENT_COUNT]; N_FLOOR as usize]
     ) -> Self {
         let clients = merged_hall_requests.into_iter()
             .zip(clients_cab_requests.into_iter())
@@ -39,9 +43,10 @@ impl FullControllerRequestsMatrix {
         }
     }
 
+    /// Convert the [FullControllerRequestsMatrix] into a raw bytes array to use in network related code.
     pub(super) fn encode(&self) -> [u8; FCRM_RAW_SIZE] {
         let mut raw_output = [0u8; FCRM_RAW_SIZE];
-        raw_output[0..CLIENT_COUNT as usize].copy_from_slice(&self.client_ids);
+        raw_output[0..CLIENT_COUNT].copy_from_slice(&self.client_ids);
         let clients: [u8; CLIENTS_RAW_SIZE] = self.clients
             .iter()
             .flatten()
@@ -49,16 +54,17 @@ impl FullControllerRequestsMatrix {
             .collect::<Vec<u8>>()
             .try_into()
             .unwrap();
-        raw_output[CLIENT_COUNT as usize..FCRM_RAW_SIZE].copy_from_slice(&clients);
+        raw_output[CLIENT_COUNT..FCRM_RAW_SIZE].copy_from_slice(&clients);
         raw_output
     }
 
+    /// Inverse function of [encode](FullControllerRequestsMatrix::encode)
     pub(super) fn decode(raw_bytes: &[u8]) -> Self {
         debug_assert_eq!(raw_bytes.len(), FCRM_RAW_SIZE);
-        let mut client_ids = [0u8; CLIENT_COUNT as usize];
-        client_ids.copy_from_slice(&raw_bytes[0..CLIENT_COUNT as usize]);
+        let mut client_ids = [0u8; CLIENT_COUNT];
+        client_ids.copy_from_slice(&raw_bytes[0..CLIENT_COUNT]);
 
-        let raw_clients: [bool; CLIENTS_RAW_SIZE] = raw_bytes[CLIENT_COUNT as usize..]
+        let raw_clients: [bool; CLIENTS_RAW_SIZE] = raw_bytes[CLIENT_COUNT..]
             .iter()
             .map(|x| *x != 0)
             .collect::<Vec<bool>>()
@@ -79,6 +85,11 @@ impl FullControllerRequestsMatrix {
         }
     }
 
+    /// Get the requests matrix of a specific elevator.
+    /// This matrix contains both the hall requests and cab requests.
+    ///
+    /// Since synchronisation doesn't reschedule call on the Backup, we can store all hall requests
+    /// inside the first elevator and only store the respective cab requests in every elevator.
     pub fn get_requests_matrix_of(
         &self,
         elevator_id: ConnectionIdentifier,
@@ -95,6 +106,10 @@ impl FullControllerRequestsMatrix {
             })
     }
 
+    /// Get the cab requests part of the requests matrix for a specific elevator.
+    ///
+    /// Since synchronisation doesn't reschedule call on the Backup, we can store all hall requests
+    /// inside the first elevator and only store the respective cab requests in every elevator.
     pub fn get_cab_requests_of(
         &self,
         elevator_id: ConnectionIdentifier
